requirements.txt
asgiref==3.2.7
astroid==2.3.3
biopython==1.76
certifi==2019.11.28
colorama==0.4.3
Django==3.0.4
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
# mkl-fft==1.0.15
# mkl-random==1.1.0
# mkl-service==2.3.0
# numpy==1.18.1
Pillow==7.1.1
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.15
pylint-mongoengine==0.4.0
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wincertstore==0.2
wrapt==1.11.2
